imporve finding user location and seting pins on map

+++ Map/Map/Model/Pin.swift
@@ -11,12 +11,13 @@ import Foundation
 final class Pin {
     
     let id: Int
-    let lng: Float
-    let lat: Float
+    let lng: Double
+    let lat: Double
     let name: String
     let imagePath: String
     
-    init(id: Int, lng: Float, lat: Float, name: String, imagePath: String) {
+    //MARK:- TODO usunąć
+    init(id: Int, lng: Double, lat: Double, name: String, imagePath: String) {
         self.id = id
         self.lng = lng
         self.lat = lat
@@ -27,8 +28,8 @@ final class Pin {
     init(withDictionar pinDict: [String:Any])   {
         id = pinDict[APIKeys.id] as? Int ?? 0
         name = pinDict[APIKeys.name] as? String ?? ""
-        lat = pinDict[APIKeys.lat] as? Float ?? 0
-        lng = pinDict[APIKeys.lng] as? Float ?? 0
+        lat = pinDict[APIKeys.lat] as? Double ?? 0
+        lng = pinDict[APIKeys.lng] as? Double ?? 0
         imagePath = pinDict[APIKeys.avatar] as? String ?? ""
     }    
 }
