+++ Map/Map/Utilities/URLBuilder.swift
@@ -9,7 +9,7 @@ import Foundation
 
 final class URLBuilder {
 
-    static func pinsURL(withLat lat: Float, withLng lng: Float) -> URL {
+    static func pinsURL(withLat lat: Double, withLng lng: Double) -> URL {
         let urlString = "https://interview-ready4s.herokuapp.com/geo?lat=\(String(lat))&lng=\(String(lng))"
         return URL(string: urlString)!
     }
